Fix for 2787 when using nrepl-hide-special-buffers
nrepl-hide-special-buffers would prepend a leading space to the nrepl
process buffer to hide it from the buffer list. If you created a new
sibling process there was a bug however. In order to uniquely name
buffers, CIDER uses `generate-new-buffer-name` to ensure a buffer is
uniquely named. However, the space was put on _after_ this
call. The sequence is as follows:

- jack in
- use template "*cider-nrepl blah" to see if existing process buffer
is around. Its not. Since we want these hidden, add a leading space so
now " *cider-nrepl blah*" process buffer exists
- watch for "nREPL server started on port \\([0-9]+\\)" and not
already a port saved
- start repl
- jack in again creating sibling process
- check if buffer "*cider-nrepl blah" exists. It doesn't since the
first one has a leading space
- add a leading space to the buffer name to get " *cider-nrepl blah"
which actually now is a collision
- lein/deps/shadow startup output goes in the same buffer as the other
process and therefore no new repl is started up when watching the the
process output since the following code is waiting:

```lisp
(when (and (null nrepl-endpoint)
           (string-match "nREPL server started on port \\([0-9]+\\)" output))
```