[Fix #2527] Enable configuring REPL auto-clear
Enables setting a limit to the buffer size of the REPL buffer (off by
default), which when exceeded after printing out a result or other
output, will clear old REPL output from the beginning to ensure it stays
within the limit. When clearing it also ensures partial input or output
is erased.