Translate paths from CIDER to nREPL and vice-versa
When `cider-path-translations` is in use (e.g., when CIDER is running
on the host, but the nREPL is running in a virtual machine or a
container) we need to translate from the host "CIDER-based" paths to
the container "nREPL-based" paths and vice-versa, depending on the
operation.

Translations from "nREPL-based" paths to "CIDER-based" paths were
already implemented, but not the other way around. Some operations in
clj-refactor need the "CIDER-based"to "nREPL-based" path translations
to work correctly (e.g., `cljr-rename-file-or-dir`,
`cljr-rename-symbol`, etc.).

This patch implements the missing path translations, and is a
pre-requisite for the complementary clj-refactor patch that is
proposed separately.