Fix format of clojure-cli options because TRAMP
This is /a/ solution for https://github.com/clojure-emacs/clj-refactor.el/pull/445#issuecomment-628206640. I do not know what the cause of this problem is. Perhaps `IFS` is different between cider-jack-in versus directly on the command-line? Perhaps it's because TRAMP uses `/bin/sh` to run remote commands, not `/bin/bash`? I don't know. Given that `,` is whitespace in clojure, this seems like a harmless patch.