Reintroduce main opts for jvm options
these jvm options need to go at the beginning. Once the cli tools are
updated for `-A` to not use main opts all of this goes away and
everyone can just use the `cider-clojure-cli-global-options`. But
until that is the case people will need to use
`cider-clojure-cli-aliases` if they have aliases with mains.