Provide ability to customise cljfmt code formatting
**Describe the solution you'd like**

It should be possible to provide a custom formatting configuration for
cider-nrepl/cljfmt. The cljfmt configuration format is described here:

  https://github.com/weavejester/cljfmt#configuration

The desired solution should have a defcustom variable where the
configuration can be provided.

**Describe alternatives you've considered**

A workaround to this solution is to use an after-save-hook as
described here:

https://github.com/dzer6/cljfmt-graalvm

However, in this case, the two formatters (cider, external cljfmt)
will keep fighting each other causing an awful user experience.

**Additional context**

In https://github.com/clojure-emacs/cider/issues/2907 Bozhidar Batsov
commented:

 > You can just disable dynamic (REPL-powered) indentation if you
 > don't rely on it
 > https://docs.cider.mx/cider/0.26/config/indentation.html
 > Alternative you can completely override the indentation logic of
 > clojure-mode and provide whatever implementation you want. CIDER
 > simply hooks into it and adds new indentation rules there
 > dynamically when connected.

To which Brunno Bonacci responded:

  > However, I still think that this ticket stands. Having the
  > possibility to configure the cljfmt indentation rules is a
  > nice-to-have feature.