Use appropriate coding system when unzipping jars
in `cider-find-file`. This prevents CRLF encoded files from appearing with ^M
characters at the end of each line, which, in turn, avoids confusing
`clojure-find-ns` and breaking `cider-find-var`.
`clojure-find-ns` uses Emacs' `(thing-at-point 'symbol)` as part of identifying
a file's namespace, and when a file isn't decoded properly, namespaces can be
reported as `my.lib^M` which `cider-find-var` won't know what to do with.