[Fix #3020] Session linking on Windows
When jumping into a file on the classpath it isn't properly linked with
the current buffer at present.  This commit fixes that, by normalizing
the file paths returned by `cider-classpath-entries`.

Without this normalization step the comparison in
`sesman-friendly-session-p` returns the wrong results when comparing the
current buffer with entries on the class path.

The normalized path will look like this `c:/myFile` whereas the raw one,
on Windows, will look like this: `c:\\myFile`.