Sideloader: no line breaks in base64 content
Currently we are sending sideloaded resources as line-wrapped base64. The extra
newlines confuse the nREPL sideloader, turning the result into binary garbage.

I think in this case we should be both more conservative in what we send (this
patch) and more lenient in what we receive (nREPL should sanitize the input and
drop any characters that aren't valid base64).