Allow starting the sideloader for the tooling session
CIDER's deferred loading of middleware means that requires can happen during all
kinds of nREPL ops, many of which are performed over the tooling session. To
make this work via the sideloader we need it to be active both on the regular
session and on the tooling session, so provide a flag for starting it in either
session.