Allow fallback to xref backends if cider-nrepl is unloaded
This commit addresses the following use-case: Connecting to a running
REPL which does not have `cider-nrepl` middleware loaded.

This can happen when connecting to remote REPLs, and even
locally (admittedly minor use-case) if you don't want to turn the
middleware on.

If `cider--xref-backend` returns nil, `xref` will move on to the next
backend. This allows using something like `dumb-jump` to move around
in the code-base, while also getting the benefits of a REPL window.