clojure-cli: provide a mechanism to configure default aliases
The clojure CLI does not impose any convention on what the default
alias for a project should be. It is also common practice to
only have the necessary environment for production in the
root dependency expressions.

As a result, jacking in without modification often results in
a REPL environment that is too constrained.

This patch looks up the current `deps.edn` file to look for a
`:cider/default-aliases` key, when that key is present and
contains a collection, it is processed to produce the correct
alias stanza for the clojure CLI commandline.