Consecutive result overlays can now seen
Previously, if multiple overlays were created one after another that
are meant to last for one command, every second overlay wouldn't show
up. This is because there's a hook in the 'post-command-hook that will
delete all overlays, and it runs right after the second overlay is
created.

Instead, if we're creating a new 'result overlay as we speak (ie. we
are in the function 'cider--make-result-overlay), we should call
'cider--remove-result-overlay directly, so that the previous overlay
gets removed AND 'cider--remove-result-overlay  gets removed from the
'post-command-hook.

We can conceptualize this as running the hook early, since we're
deciding to delete the overlay now rather than wait till the
'post-command-hook.