For Gradle jack-in use wrapper by default
While Gradle does support a global install on the PATH called via
"gradle" it is much more common for projects to have the Gradle wrapper
in their project, which allows each project to be pinned to a specific
version of Gradle.

This file is either "gradlew" or "gradlew.bat" in the project root.

Since this file isn't on the exec-path, I use locate-file to resolve the
file relative to the clojure-project-dir, but only if the file is
clearly a relative path starting "./" or "../". We still fall back to
looking up on the exec-path if someone sets their cider-gradle-command
back to "gradle".