Don't disable Gradle daemon by default
When Gradle support was originally added to cider in 2015, the default
was set to disable Gradle's daemon [1] which Gradle uses to keep a warm
JVM around to reuse for future runs.

While I know there are some strong opinions about its use of a daemon,
this is the default behavior of Gradle and I don't think we have any
strong reasons to disable it in this case.

It's possible in 2015 the daemon was less stable or more of a hog, but I
don't believe that's true nowadays.

[1] https://docs.gradle.org/current/userguide/gradle_daemon.html