Consecutive result overlays can now be seen (#3259)
Previously, if multiple overlays were created one after another that
are meant to last for one command, every second overlay wouldn't show
up.

This is because there's some bad hook management causing the second
overlay in a sequence to be deleted. Whenever we're creating a second
overlay, the first one is automatically removed. But,
`cider--remove-result-overlay' is still in the `post-command-hook',
and thus, right after the second overlay is created, it's immediately
deleted.

Instead, when we're creating a new overlay, we should call remove any
instances of `cider--remove-result-overlay' in the
`post-command-hook', so that the newly created overlay doesn't get
hosed.