Don't scale down time for overlay tests on MacOS
Overlay tests involve testing the timer functionality of overlays, and
require the use of sleep. The tests scale all time-related arguments
by a certain factor to avoid overly inflating test-running
times.

The scaling factor was initially set at 1 order of magnitude slower
than the consistently succeeding value on my machine, but the value is
apparently still too fast for the MacOS test runner, so I have added
MacOS to the list of platforms unsafe to do this optimization for.