Move helper functions to own file, fix test issue with client conn
Changes
1. Created integration-test-utils.el to collect increasing number of
integration helper functions.
1. A `with-cider-test-sandbox` macro is introduced to restore cider
state after test has finished.
2. The `cider-itu-poll-until` helper function is introduced to replace
`nrepl-tests-sleep-until`, this fn now throws if condition does not
become true after the timeout has elapsed.
3. There was an issue with the previous attempt to add a hook and
return a value for the test to poll when he client is connected to
the nREPL. The hook was naively attached to the temp buffer, which is
not necessarily in scope when the hook is run. A new
`cider-itu-nrepl-client-connected-ref-make!` replaces that fn, that
now returns a global variable ref, and has to be called inside
`with-cider-test-sandbox` so that the hook variable is restored after
the test finishes.
4. Introduced tests to test the new integration tests helper functions.
5. All integration tests are updated to use the new functions mentioned above.
6. Rename `connected?` to `is-connected` as per review comments.
7. Restore `nrepl-client-tests.el` update which was overwritten with
previous commit.