Add nREPL server logging, inc time out, don't kill server buf on err
This is a temporary hack for the Clojure cli and shadow integration
tests, which intermittently timing out on eval (see #3298), to capture
and dump out the input and output traffic on the nREPL server for
post-mortem investigation. The time out period is also increased to
30secs.

In addition to this, the `nrepl-server-sentinel' does not kill the
`server-buffer` when there is a startup error, since the buffer can
hold valuable information for the user to inspect what might have gone
wrong during start up (this is probably a regression from the
integration tests restructuring).