Fix #3307: Incorrect special-form eldocs
This fix is the CIDER companion to [this
PR](https://github.com/clojure-emacs/orchard/pull/166) in orchard,
which lets us know if the symbol we're querying eldocs for is a
special form or an ordinary function.

This changes uses that information to deduplicate the arglists for the
special form. In the process, the change adds a
`cider-eldoc-format-special-form` function, which serves as a home for
special form formatting.

This change also adds a unit test for cider-eldoc to ensure that it
calls the appropriate formatting methods.