Modified `xref-backend-references` to use buffer locations instead of filenames
The filenames returned by `cider-sync-request:fn-refs` were not in the format expected by uxref,
probably because they are relative to the project root and not relative to the current buffer (which
is what xref needs, from my understanding). I've implemented a somewhat dirty fix by using identical code to `xref-backend-definitions`, which uses `cider--find-buffer-for-filename`. This function seems to work around the issue, with the drawback of potentially opening a large amount of buffers.

The cleaner way to fix this bug would be computing absolute file names, but I am unsure what the filenames returned by `cider-sync-request:fn-refs` are relative to.