Add preliminary support for clojure-ts-mode
This does NOT intended to break the dependency from cider to
clojure-mode. It is intended to make CIDER work with clojure-ts-mode.

Some functionality like clojure-find-ns, clojulore-find-def, etc will
still require clojure-mode in order for CIDER to get it's work done.

Adds util functions for checking if buffer is managed by a clojure mode