Fix #3541: cider-jack-in failing with ssh-remotes
Calling cider-jack-in from a tramp buffer on a ssh-remote would throw:
"error in process filter: Wrong type argument: stringp, nil"

This happens because nrepl--ssh-tunnel-connect is trying to infer the
hostname from the current filename: It string-matches
on (buffer-file-name) which returns nil in
the *cider-uninitialized-repl* buffer.

For the jack-in use-case, the local var nrepl-project-dir will
always be bound when connecting the client, so we can use
it as a fallback.

Fixes:
https://github.com/clojure-emacs/cider/issues/3541
https://github.com/clojure-emacs/cider/issues/3303 (partially)