Add cider-log-show-frameworks command
This command renders the available log frameworks in a buffer. It
shows the log framework name, the website and javadocs urls and the
levels for now.

This is mostly to address @vemv's suggestion to ask the user to run
`(cider-sync-request:log-frameworks)`. I hooked the command up at the
usual places and added it with a small troubleshooting section to the
docs.

We could get even more sophisticated and show the current appenders
and consumers but I need a designer for this. Let's do this at some
other point.